+++ totp/workers/bt_type_code/bt_type_code.h
@@ -27,7 +27,7 @@ void totp_bt_type_code_worker_free(TotpBtTypeCodeWorkerContext* context);
 void totp_bt_type_code_worker_start(
     TotpBtTypeCodeWorkerContext* context,
     char* code_buf,
-    size_t code_buf_length,
+    uint8_t code_buf_length,
     FuriMutex* code_buf_update_sync);
 void totp_bt_type_code_worker_stop(TotpBtTypeCodeWorkerContext* context);
 void totp_bt_type_code_worker_notify(
