+++ totp/workers/usb_type_code/usb_type_code.h
@@ -22,7 +22,7 @@ enum TotpUsbTypeCodeWorkerEvents {
 
 TotpUsbTypeCodeWorkerContext* totp_usb_type_code_worker_start(
     char* code_buf,
-    size_t code_buf_length,
+    uint8_t code_buf_length,
     FuriMutex* code_buf_update_sync);
 void totp_usb_type_code_worker_stop(TotpUsbTypeCodeWorkerContext* context);
 void totp_usb_type_code_worker_notify(
